Create Range Partition.java
The basic approach is to use three index. First left, second right and third to traverse the array. Index left starts form 0, Index right starts from N-1. We traverse the array whenever we find a value lower than range we swap it with the value at start and increment start. And whenever we finds a value greater than range we swaps this value with right and decrement right. When traversal is complete we have the array partitioned about range.